lesson6/pyproject.toml
[project]
name = "lesson6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "tkeasygui>=1.0.40",
]
